Revert "Feat/prisma pokemon ability moves schema"
