--- Dockerfile
@@ -1,15 +1,0 @@
-FROM node:22-slim
-
-WORKDIR /app
-
-RUN apt-get update -y \
- && apt-get install -y openssl \
- && rm -rf /var/lib/apt/lists/*
-
-COPY package*.json ./
-RUN npm ci
-
-COPY prisma ./prisma
-RUN npx prisma generate
-
-COPY . .
