--- docker-compose.yml
@@ -1,43 +1,0 @@
-services:
-  db:
-    image: postgres:15
-    container_name: pokemon-db
-    environment:
-      POSTGRES_USER: ${DB_USER}
-      POSTGRES_PASSWORD: ${DB_PASSWORD}
-      POSTGRES_DB: ${DB_NAME}
-    ports:
-      - "${DB_PORT}:5432"
-    volumes:
-      - postgres_data:/var/lib/postgresql/data
-    healthcheck:
-      test: ["CMD-SHELL", "pg_isready -U ${DB_USER} -d ${DB_NAME}"]
-      interval: 5s
-      timeout: 3s
-      retries: 10
-  
-  backend:
-    build: 
-      context: .
-      dockerfile: Dockerfile
-    command: npm run start:dev
-    container_name: pokemon-backend
-    env_file:
-      - .env
-    ports:
-      - "${BACKEND_PORT}:${BACKEND_PORT}"
-    depends_on:
-      db:
-        condition: service_healthy
-        
-    environment:
-      DB_HOST: db
-      CHOKIDAR_USEPOLLING: "true"
-      PORT: ${BACKEND_PORT}
-    
-    volumes:
-      - .:/app             
-      - /app/node_modules
-
-volumes:
-  postgres_data:
